Lint for headings in text that aren’t in the ToC
This collects text from the headings that has had the tags stripped and whitespace normalised. It then checks them off against the ToC entries to see if any are present that don’t have an entry.

The only amend we do to the ToC entries is to strip out the ‘:’ that the style guide suggests for after the chapter number.

Tested against the SE corpus.