Don’t lint missing subtitles in ToC when they don’t follow a roman numeral
BS was really messing me around here, but I think I’ve landed on something that’s correct.

This checks to see that a roman numeral span is the first span in a heading: if it’s not then the subtitle is removed from the expected ToC entry.

With this patch Chekov’s short fiction passes linting.