Fix endnote reordering
In the new endnote style the links to the notes have an `id` attribute starting with `noteref-`, but the existing code is still looking for `note-`.

Obviously this commit will break reordering endnotes in the old style, so if older books have endnotes added they’ll need to be updated to the new style first.