Check the spine’s order against the ToC order
Add a spine check to the ToC section of the linter. Previously we were getting all ToC entries with find_all('a') but we need them depth first to match the spine order. This first changes the heading selection order to be in (hopefully) spine order, then checks each against the corresponding spine entry for a match. If it fails it throws an single failing error for the whole spine and breaks out.