Handle spine lining case where a ToC only references anchors in that file
Previously we were naively throwing away any links to an anchor, but there are some books where the only reference to a file in the ToC is with anchor links. This updates the code to collect all references to a file, strip all anchors, and then pull out only unique entries from that list.