Make sure LoI text / image title / figure alt text are consistent
The title text in the SVG should always match the alt text in the associated figure, so that’s easy to lint and error on (if there’s no title, eg a PNG we can bail). As well as that, the title text should mostly match the text in the LoI but there are cases where this doesn’t work (eg Just So Stories), so if it doesn’t we emit a manual review warning.