Add a warning if figcaption doesn’t match LoI text
A warning and not an error as it’s possible (e.g. Just So Stories) for figcaptions to differ significantly from LoI text on purpose, although this generally isn’t the case. This also removes the warning that matched SVG `<title>`s against LoI text as that turned out to be problematic.