Warn on >> at the end of a tag
Seen this a few times, so it’s worth linting for. We also need to check for >&gt; as clean will ‘fix’ the second > if run before lint, and it isn’t caught in a manual review.

We don’t need to check for double opening < as that throws a parse error in our tooling.