Stop looking for more endnotes if we’ve made some changes and run out
The original loop is naïve and will keep on looking for endnotes in a situation where there definitely won’t be any (assuming that the endnote structure is correct). If we’ve already found one or more endnotes and made changes then if we fail to find the next in the loop we can assume that the rest don’t exist. Against current Pepys and incrementing all endnotes by one this commit drops the total time from 31.5s to 14.5s(!).