Speed up note link text replacement
The most expensive part of the replacement process is the regular expression for the link. But we know that the endnote link will just contain a number, and will be well-formed. Consequently, we can use a simple text replacement instead of an expensive regex.

This drops the average speed to increment every endnote on current Pepys from 20.2s to 5.7s.