Fix stripping of abbrs with classes when file has a MathML namespace
LXML is including namespace information that we normally strip, but it doesn’t strip the MathML namespace when it’s used. I haven’t analysed the code fully, but not doing this leads to the abbr not being replaced by a span and consequently losing any applied styling through the class.