Handle more complicated pseudoclass arrangements in epub2 rewriting
Fixes a bug in the existing code where, given a selector like `p:first-child + p`, the second p would gave a `first-child` class instead of the first. We now remove the trailing part to make the element targeted by the pseudoclass the element that gains the class.

This also now handles a situation where there are multiple pseudoclasses in a single selector, by looping until there are no more classes to add. I don’t think this is particularly useful given our current set of selectors to simplify (`p:first-child:last-child` can be replaced by `p:only-child`) but it might stop an unlucky scenario, and we may change SELECTORS_TO_SIMPLIFY in the future.