Fix extra gap on Kobo before 4-dot ellipses
The Kobo build-chain breaks sentences into “kobo spans”. These shouldn’t affect the rendering, but as it stands a hair space in the content becomes a full space in the presentation.

This is most obvious with 4 dot ellipses, which we replicate with a full stop followed by a hair space followed by an ellipsis. This revised regex looks for that pattern and avoids splitting, causing the 4-dot ellipsis to render as expected.