Fix missing generated epub-type-x classes
The code that generates classes to bypass spotty pseudoclass support was expecting to always run first, before the code that added generated epub-type classes. For reasons unknown, this is sometimes happening in reverse. When that happens, the code that generates pseudoclass replacements simply overwrites the now-earlier epub-type classes.

While a correct fix would be to structure the code to make it deterministic, we can at least work around it by backporting (and hence normalising) the class-setting logic from the epub-type routine to the pseudoclass routine.

With this applied we still see evidence of the non-determinism, but it’s no longer destructive (for example, we could see both `class="first-child epub-type-z3998-persona"` and `class="epub-type-z3998-persona first-child"`).