Improve url generation
The existing solution transposed characters like ö to o, but failed on characters like ø. We can instead use the Unidecode library, which covers automatically everything we’d want to theoretically do.

This adjusts create-draft and formatting to use the new library. It doesn’t change find-mismatched-diacritics, as identifying what a “diacritic” is here is a harder problem.