Update spelling of Kyiv
I hadn’t realised this, but this has been the Ukrainian preferred spelling since ’95 and is a UN standard, with a campaign to get foreign media to change starting in 2018: https://en.wikipedia.org/wiki/KyivNotKiev. Given the recent standardisation of the media on “Kyiv” (driven by the current invasion) it feels right to also update our corpus to the more modern spelling.