Add a check for a quote starting with a lowercase character after a period
It was slightly surprising to me that we aren’t already checking for this, but an easy thing to add. There are currently 33 instances of this in my clone of the corpus, but each would need to be double-checked against the source files to see whether it’s a problem with the period or the lowercase character.