create-draft: Don't copy metadata from template files
On NixOS, the se package (including templates) gets installed to a
read-only filesystem. Using shutil.copy on the templates copies the
read-only mode! shutil.copyfile avoids copying the metadata, but
requires that the destination is a path to a file (not a directory),
which we now construct in the same way as shutil.copy does.