Fix cover building as only a black image
Looks like cairosvg 2.7.0 broke loading data URLs in source SVGs in “safe mode”, which causes the produced PNG to have a transparent background and the rendered JPG to end up with an all-black background. This was fixed in April (https://github.com/Kozea/CairoSVG/commit/2cbe3066e604af67c31d6651aa3acafe4ae0749d) but the latest 2.7.0 release was in March.

In the meantime, we can pass an unsafe flag through to svg2png which allows it to complete successfully. I don’t think this is a problem as we control the repos that build this data.