Lowercase Arabic definite article in the middle of a sentence
This is not the most fun regex. We use a negative lookbehind to check that we’re not following a sentence ending character (full stop, exclamation mark or question mark). The following space check also makes sure that we’re in the middle of a sentence. Then we check after the `Al-` for a capital letter to try to avoid selecting for e.g. `Al-bert`.

There will be some missed opportunities here. The biggest problem will constructions like `<i xml:lang="ar-Latn">Al-Zubayr` that won’t be picked up. But there’s a limit to how complex I want to make the regex.