Make the reading ease score cope better with free-form poetry
When processing the text to calculate the reading ease score, we can first preprocess it to allow for verse that doesn’t end with punctuation. If we add a full stop at the end of verses just for the reading ease algorithm then we get more normal scores, and it shouldn’t affect other productions in a meaningful way.

1. Mina Loy’s Poetry goes from -128.58 to 42.45
2. William Carlos William’s Poetry goes from 79.5 to 79.6
3. Laurence Sterne’s Tristram Shandy goes from 51.35 to 51.55
4. Every other repo I’ve tried has no change.