Fix wrong click behaviour in Apple Books with drama
Apple Books has a specific behaviour for tables: if you click on them you get a scannable zoomable popup of the whole table. This is really nice functionality, until you realise that we use tables to lay drama out.

A fix is to block events on drama tables. Obviously we have some links (for example endnotes) inside drama that need to remain clickable, so we unblock for those.