Testing: Move build to its own group, add controller and test
The build command is unique in what it produces and where it produces it, and does not fit within the structure of any of the
existing test groups. This breaks it out into its own group, adds a test_build_command.py controller, and adds a helper function
specific to the build command. It also adds a test for the build commands.