Testing: Move recompose-epub to stdout commands, handle stdout special cases
A few of the stdout commands have special requirements. The build-manifest/spine/toc commands can either update existing files or
output to stdout. Thus, to be tested in the stdout group, they require the --stdout parameter. The unicode_names and help commands
do not take an ebook directory as input. This adds the code necessary to accommodate both of those special circumstances.