Testing: Add build-* commands to draft group, check --stdout arg in draft/ebook
The build-loi/manifest/spine/title/toc commands can either update existing files or output to stdout. Add them to the draft group
so they can be tested either way: to stdout in draft, to files in ebook. In the draft controller, require the --stdout argument
for them in any command file; in the ebook controller, disallow the --stdout argument.