Testing: Modify files_are_golden helper function to work correctly with all callers.
The draft and ebook commands groups use the input files as the *list* of files to compare, but the results files to do the
actual comparison. The file command group uses the result files for both. This modifies the files_are_golden helper function
to accommodate both and also handles comparisons to image files, and modifies the callers to pass the new parameter.